# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for the research monitor (arXiv ingestion + paper catalogue).",
        "title": "Research Monitor API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/analyses": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyses"
                ],
                "summary": "Submit an analysis",
                "parameters": [
                    {
                        "description": "Analysis request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_analyzer.SubmitAnalysisRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analysis persisted",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_analyzer.AnalysisEnvelope"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Extraction not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Extraction not in done status",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Analysis storage unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "502": {
                        "description": "LLM upstream failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/analyses/{extraction_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyses"
                ],
                "summary": "Get an analysis by extraction id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Extraction id",
                        "name": "extraction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analysis found",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_analyzer.AnalysisEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Analysis not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Analysis storage unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/arxiv/fetch": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Triggers an arXiv fetch + persist cycle. No body, no query params.\nReturns the fetched entries with per-entry is_new derived from\nwhether persistence inserted a new row for that paper.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Arxiv"
                ],
                "summary": "Fetch latest arXiv entries",
                "responses": {
                    "200": {
                        "description": "Fetched entries",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_arxiv.FetchEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Catalogue unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "502": {
                        "description": "Upstream arXiv returned a non-success or malformed response",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "504": {
                        "description": "Upstream arXiv unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/extractions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Extractions"
                ],
                "summary": "Submit an extraction",
                "parameters": [
                    {
                        "description": "Extraction request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_extraction.SubmitExtractionRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Extraction enqueued",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_extraction.ExtractionStatusEnvelope"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Catalogue unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/extractions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Extractions"
                ],
                "summary": "Get extraction status by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Extraction id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Extraction state",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_extraction.ExtractionStatusEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Extraction not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Catalogue unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_yoavweber_research-monitor_backend_internal_http_common.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.Error"
                }
            }
        },
        "github_com_yoavweber_research-monitor_backend_internal_http_controller_paper.DownloadEntryResultDTO": {
            "type": "object",
            "properties": {
                "bytes": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "paper_id": {
                    "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_controller_paper.PaperIDDTO"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_yoavweber_research-monitor_backend_internal_http_controller_paper.DownloadJobSnapshotDTO": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "completed_at": {
                    "type": "string"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_controller_paper.DownloadEntryResultDTO"
                    }
                },
                "failed": {
                    "type": "integer"
                },
                "job_id": {
                    "type": "string"
                },
                "succeeded": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_yoavweber_research-monitor_backend_internal_http_controller_paper.PaperIDDTO": {
            "type": "object",
            "properties": {
                "source": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "internal_http_controller_analyzer.AnalysisEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_http_controller_analyzer.AnalysisResponse"
                }
            }
        },
        "internal_http_controller_analyzer.AnalysisResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "extraction_id": {
                    "type": "string"
                },
                "long_summary": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "prompt_version": {
                    "type": "string"
                },
                "short_summary": {
                    "type": "string"
                },
                "thesis_angle_flag": {
                    "type": "boolean"
                },
                "thesis_angle_rationale": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_http_controller_analyzer.SubmitAnalysisRequest": {
            "type": "object",
            "required": [
                "extraction_id"
            ],
            "properties": {
                "extraction_id": {
                    "type": "string"
                }
            }
        },
        "internal_http_controller_arxiv.EntryResponse": {
            "type": "object",
            "properties": {
                "abs_url": {
                    "type": "string"
                },
                "abstract": {
                    "type": "string"
                },
                "authors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "is_new": {
                    "type": "boolean"
                },
                "pdf_url": {
                    "type": "string"
                },
                "primary_category": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                },
                "submitted_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "internal_http_controller_arxiv.FetchEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_http_controller_arxiv.FetchResponse"
                }
            }
        },
        "internal_http_controller_arxiv.FetchResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_http_controller_arxiv.EntryResponse"
                    }
                },
                "fetched_at": {
                    "type": "string"
                },
                "job": {
                    "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_controller_paper.DownloadJobSnapshotDTO"
                }
            }
        },
        "internal_http_controller_extraction.ExtractionStatusEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_http_controller_extraction.ExtractionStatusResponse"
                }
            }
        },
        "internal_http_controller_extraction.ExtractionStatusResponse": {
            "type": "object",
            "properties": {
                "body_markdown": {
                    "type": "string"
                },
                "failure_message": {
                    "type": "string"
                },
                "failure_reason": {
                    "description": "Failure branch — populated iff status == failed.",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/internal_http_controller_extraction.MetadataDTO"
                },
                "source_id": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "description": "Artifact branch — populated iff status == done. Metadata is a pointer\nso the entire object disappears from JSON when nil.",
                    "type": "string"
                }
            }
        },
        "internal_http_controller_extraction.MetadataDTO": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "word_count": {
                    "type": "integer"
                }
            }
        },
        "internal_http_controller_extraction.SubmitExtractionRequest": {
            "type": "object",
            "required": [
                "pdf_path",
                "source_id",
                "source_type"
            ],
            "properties": {
                "pdf_path": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT issued by POST /auth/login. Supply as \"Bearer \u0026lt;token\u0026gt;\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}